fix(net,linux): remove dep on libanl
